requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.1.78
firebase-admin==6.2.0
python-dotenv==1.0.0
